Add -H:+AddAllCharsets option to be safe.
I got an error about not having the Cp1252 charset at one point,
but it disappeared after adding reflection config for the Bean
classes. Yet I can't reproduce it now. Since I don't know the cause,
I feel that it's best to add this option. The only penalty is an
increase in binary size. If that's a problem for your project, you can
remove it and take your chances.